Updated run_fredi
Updated the `run_fredi()` function with the new method for calculating SLR impacts for SLR heights that fall above the defined range. Required making some adjustments to the calculation of GCM impacts (column names). The earlier update to `utils.R` also should have mentioned the creation of the function `get_scenario_id()` used to facilitate the standardization of the creation of scenario IDs used in matching GCM impact functions to scalars. Note: in the future, `createSystemData()` also uses scenario IDs, and this function can be added into that routine.